medicare2/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
beanie==1.23.0 